# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rail Yard API
  version: 1.0.0
  description: |
    Rail Yard exposes control-plane, worker, and P5 operations APIs.

    Acceptance and canonical terminal outcomes are durable and idempotent.
    Payload execution is at least once. A worker can complete an external side
    effect and lose its lease before reporting completion, so a later attempt
    can execute the payload again. The lease `idempotency_key` must be honored
    by external payloads that require idempotent side effects.
  x-payload-execution-semantics: at-least-once
...
servers:
  - url: /
...
paths:
...
  /v1/jobs:
    post:
      tags: [Jobs]
      operationId: submitJob
      summary: Submit one logical job
      description: |
        The idempotency key is scoped to the normalized job tenant. Reusing the
        key with the same normalized request returns the original response.
      x-payload-execution-semantics: at-least-once
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitJobRequest"
      responses:
        "201":
          description: The job was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitJobResponse"
        "200":
          description: The original idempotent response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "429":
          $ref: "#/components/responses/QueueFull"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/dead-letters/{job_id}/redrive:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      tags: [Dead letters]
      operationId: redriveDeadLetter
      summary: Create a new job from a dead letter
      description: Redrive creates a new job and never reopens the terminal job.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: A replacement job was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedriveDeadLetterResponse"
        "200":
          description: The original idempotent response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedriveDeadLetterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/QueueFull"
        "501":
          $ref: "#/components/responses/NotSupported"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/operations/jobs:
    post:
      tags: [P5 Operations]
      operationId: operationsSubmitJob
      summary: Submit a job through the operations facade
      description: |
        This mountable facade uses the same durable, tenant-scoped idempotency
        semantics as the control API.
      x-payload-execution-semantics: at-least-once
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitJobRequest"
      responses:
        "201":
          description: The job was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitJobResponse"
        "200":
          description: The original idempotent response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitJobResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "429":
          $ref: "#/components/responses/QueueFull"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/dags:
    post:
      tags: [P5 Operations]
      operationId: operationsSubmitDAG
      summary: Submit a durable DAG
      description: The DAG, jobs, and immutable edges are accepted atomically.
      x-payload-execution-semantics: at-least-once
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitWorkflowRequest"
      responses:
        "201":
          description: The DAG was durably accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitDAGResponse"
        "200":
          description: The original idempotent response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitDAGResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "429":
          $ref: "#/components/responses/QueueFull"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/jobs/{job_id}:
    parameters:
      - $ref: "#/components/parameters/JobID"
    get:
      tags: [P5 Operations]
      operationId: operationsGetJob
      summary: Get current job status
      responses:
        "200":
          description: Current materialized job state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/jobs/{job_id}/history:
    parameters:
      - $ref: "#/components/parameters/JobID"
      - $ref: "#/components/parameters/HistoryLimit"
      - $ref: "#/components/parameters/BeforeSequence"
    get:
      tags: [P5 Operations]
      operationId: operationsGetJobHistory
      summary: Get sequence-ordered job history
      responses:
        "200":
          description: A descending page of durable job events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobHistoryPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/jobs/{job_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      tags: [P5 Operations]
      operationId: operationsCancelJob
      summary: Cancel a nonterminal job
      description: |
        The repository records the actor and reason in the durable history.
        Cancellation must not reopen or rewrite a terminal outcome.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelJobRequest"
      responses:
        "200":
          description: Durable action receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReceipt"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/dead-letters/{job_id}/redrive:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      tags: [P5 Operations]
      operationId: operationsRedriveDeadLetter
      summary: Redrive a dead letter as an operator
      description: A new logical job is created. The terminal source job is unchanged.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Actor"
      responses:
        "201":
          description: A replacement job was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedriveDeadLetterResponse"
        "200":
          description: The original idempotent response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedriveDeadLetterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/QueueFull"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/tenants/{tenant_id}/queues:
    parameters:
      - $ref: "#/components/parameters/TenantID"
    get:
      tags: [P5 Operations]
      operationId: operationsListTenantQueueDepth
      summary: Get per-queue depth for one tenant
      responses:
        "200":
          description: Queue depth and slot use by queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueDepthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/workers:
    get:
      tags: [P5 Operations]
      operationId: operationsListWorkerHealth
      summary: List worker health and lease age
      responses:
        "200":
          description: Worker heartbeat and active lease age snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerHealthResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/dags/{dag_id}:
    parameters:
      - $ref: "#/components/parameters/DAGID"
    get:
      tags: [P5 Operations]
      operationId: operationsGetDAG
      summary: Get DAG jobs and immutable edges
      responses:
        "200":
          description: DAG details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DAGDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/operations/jobs/{job_id}/force:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      tags: [P5 Operations]
      operationId: operationsForceJobAction
      summary: Apply a fenced operator force action
      description: |
        `release` fences the current lease and makes the job eligible for a
        successor attempt. `fail` and `dead_letter` create one canonical
        terminal outcome. Implementations must reject stale state and must not
        mutate an existing terminal outcome.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/Actor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForceJobRequest"
      responses:
        "200":
          description: Durable action receipt.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionReceipt"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Printable ASCII key, at most 256 bytes. The same key and normalized
        request return the original response. A conflicting request returns
        `idempotency_conflict`.
      schema:
        type: string
        minLength: 1
        maxLength: 256
        pattern: "^[!-~]+$"
...
    JobID:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    DAGID:
      name: dag_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Idempotency, fencing, or state conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RequestTooLarge:
      description: Request body exceeds the configured limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    QueueFull:
      description: Tenant admission depth is full.
      headers:
        Retry-After:
          description: Suggested delay in seconds.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotSupported:
      description: The configured repository does not support this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    SubmitJobRequest:
      type: object
      additionalProperties: false
      required: [job]
      properties:
        job:
          $ref: "#/components/schemas/JobSpec"
    SubmitJobResponse:
      type: object
      additionalProperties: false
      required: [job, duplicate]
      properties:
        job:
          $ref: "#/components/schemas/Job"
        duplicate:
          type: boolean
...
    RedriveDeadLetterResponse:
      type: object
      additionalProperties: false
      required: [job, duplicate]
      properties:
        job:
          $ref: "#/components/schemas/Job"
        duplicate:
          type: boolean
...